--- docs/source/modules/cli.rst
@@ -1,4 +1,0 @@
-quota_notifier.cli
-------------------
-
-.. automodule:: quota_notifier.cli
