--- docs/source/modules/email.rst
@@ -1,4 +1,0 @@
-quota_notifier.email
---------------------
-
-.. automodule:: quota_notifier.email
