+++ docs/source/modules/main.rst
@@ -1,0 +1,4 @@
+quota_notifier.main
+-------------------
+
+.. automodule:: quota_notifier.main
