+++ docs/source/overview/usage.rst
@@ -28,6 +28,6 @@ Commandline Interface
 ---------------------
 
 .. argparse::
-   :ref: quota_notifier.cli.Parser
+   :ref: quota_notifier.main.Parser
    :prog: notifier
