+++ pyproject.toml
@@ -27,7 +27,7 @@ file = "README.md"
 content-type = "text/markdown"
 
 [project.scripts]
-notifier = "quota_notifier.cli:Application.execute"
+notifier = "quota_notifier.main:Application.execute"
 
 [project.optional-dependencies]
 docs = ["sphinx==5.2.1", "sphinx-argparse==0.3.1", "sphinx-copybutton==0.5.0", "sphinx-pydantic==0.1.1", "sphinx-rtd-theme==1.0.0", "sphinx-jsonschema==1.15", ]
