+++ quota_notifier/main.py
@@ -1,4 +1,4 @@
-"""The ``cli`` module defines the application's command line interface and
+"""The ``main`` module defines the application's command line interface and
 serves as the primary entrypoint for executing the parent package.
 It is responsible for displaying help text, parsing arguments, and
 instantiating/executing the underlying application logic.
